PhD Withdrawal State
The state of this branch at the point when I withdrawn from my PhD.
I cannot guarantee that all of the programs within the BNN_work folder are working correctly, but have left them here as syntax examples.

Over the next few days I plan to type my notes of the state of the repo, the FPEA language, and the different generation script parameters/configs and generics.

Natasha